@@ -41,145 +45,131 @@
     "@material/react-ripple": "^0.11.0",
     "@material/react-select": "^0.11.0",
     "@material/react-snackbar": "^0.11.0",
+    "@material/react-switch": "^0.11.0",
     "@material/react-text-field": "^0.11.0",
     "@material/react-top-app-bar": "^0.11.0",
-    "@material/select": "^2.3.1",
-    "@types/classnames": "^2.2.7",
+    "@material/select": "^3.2.0",
+    "@n1ru4l/use-async-effect": "1.1.3",
+    "@types/classnames": "^2.2.9",
     "@types/material-components-web": "^0.43.1",
-    "@types/react": "^16.0.25",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.9",
-    "@types/react-router-dom": "^4.3.3",
-    "animate": "^1.0.0",
-    "animsition": "^4.0.2",
-    "auth0-js": "^9.10.4",
-    "axios": "^0.18.0",
-    "bootstrap": "^4.3.1",
-    "canvas": "^2.5.0",
-    "chart.js": "^2.8.0",
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
+    "@types/react-redux": "^7.1.2",
+    "@types/react-router-dom": "^4.3.5",
+    "autoprefixer": "^9.6.1",
+    "axios": "^0.19.0",
+    "browserslist": "^4.7.0",
+    "canvas": "^2.6.0",
     "classnames": "^2.2.6",
     "codecov": "^3.5.0",
     "cookies-js": "^1.2.3",
     "countries-list": "^2.4.3",
     "cross-env": "^5.2.0",
     "cssnano": "^4.1.10",
-    "dompurify": "^1.0.10",
-    "dotenv": "^5.0.1",
-    "dotenv-webpack": "^1.7.0",
-    "draft-js": "^0.10.5",
-    "draftjs-to-html": "^0.8.4",
-    "express": "^4.17.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "faker": "^4.1.0",
+    "date-fns": "2.1.0",
+    "dotenv": "^8.1.0",
+    "express": "^4.17.1",
     "file-saver": "^2.0.2",
-    "firebase": "^6.1.0",
-    "history": "^4.9.0",
-    "html2canvas": "^1.0.0-alpha.12",
-    "jest-canvas-mock": "^2.1.0",
+    "firebase": "^7.0.0",
+    "firebaseui": "^4.2.0",
+    "history": "^4.10.1",
+    "html2canvas": "^1.0.0-rc.3",
+    "jest-canvas-mock": "^2.1.1",
     "jest-localstorage-mock": "^2.4.0",
-    "jquery": "^3.4.1",
-    "jquery.easing": "^1.4.1",
-    "jspdf": "^1.5.3",
+    "jest-resolve": "^25.0.0",
     "jwt-decode": "^2.2.0",
-    "material-components-web": "^2.2.0",
-    "material-table": "^1.40.1",
     "moment": "^2.24.0",
+    "moment-timezone": "^0.5.14",
     "moxios": "^0.4.0",
     "popper.js": "^1.15.0",
     "postcss-loader": "^3.0.0",
-    "query-string": "^6.5.0",
-    "rc-progress": "^2.3.0",
-    "react": "^16.8.6",
-    "react-beautiful-dnd": "^11.0.3",
-    "react-chartjs-2": "^2.7.6",
-    "react-country-region-selector": "^1.4.3",
+    "prop-types": "^15.7.2",
+    "rc-progress": "^2.5.2",
+    "react": "^16.9.0",
+    "react-country-region-selector": "^1.4.4",
     "react-date-range": "^1.0.0-beta",
     "react-dates": "^20.2.3",
-    "react-dom": "^16.8.6",
-    "react-draft-wysiwyg": "^1.13.2",
-    "react-ga": "^2.6.0",
+    "react-dom": "^16.9.0",
+    "react-ga": "^2.7.0",
     "react-html-parser": "^2.0.2",
     "react-load-script": "^0.0.6",
-    "react-onclickoutside": "^6.8.0",
-    "react-particles-js": "^2.6.0",
-    "react-redux": "^7.0.3",
-    "react-router-dom": "^5.0.0",
-    "react-slick": "^0.24.0",
+    "react-particles-js": "^2.7.0",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.0.1",
     "react-times": "^3.1.12",
-    "react-tooltip": "^3.10.0",
-    "react-with-firebase-auth": "^1.0.3",
-    "react-wow": "^1.0.0",
-    "redux": "^4.0.1",
+    "react-with-firebase-auth": "^1.0.5",
+    "redux": "^4.0.4",
     "redux-devtools-extension": "^2.13.8",
     "redux-immutable-state-invariant": "^2.1.0",
     "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.2",
     "sass-rem": "^2.0.1",
-    "start-server-and-test": "^1.9.0",
-    "webpack": "^4.31.0"
+    "start-server-and-test": "^1.10.2",
+    "typescript": "^3.6.3",
+    "webpack-merge": "^4.2.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.4",
+    "@babel/core": "^7.6.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.4.4",
-    "@commitlint/cli": "^7.2.1",
-    "@commitlint/config-conventional": "^7.1.2",
-    "@types/enzyme": "^3.1.10",
-    "@types/jest": "^22.2.3",
-    "autoprefixer": "^9.5.1",
-    "awesome-typescript-loader": "^5.0.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^24.8.0",
+    "@babel/preset-env": "^7.6.2",
+    "@babel/preset-flow": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.6.0",
+    "@commitlint/cli": "^8.2.0",
+    "@commitlint/config-conventional": "^8.2.0",
+    "@types/enzyme": "^3.10.3",
+    "@types/jest": "^24.0.18",
+    "awesome-typescript-loader": "^5.2.1",
+    "babel-core": "^6.26.3",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-module-resolver": "^3.1.1",
+    "babel-plugin-add-module-exports": "^1.0.2",
+    "babel-plugin-module-resolver": "^3.2.0",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "chromedriver": "^2.38.3",
-    "clean-webpack-plugin": "^0.1.19",
-    "css-loader": "^0.28.11",
-    "cypress": "^3.1.5",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.6.0",
-    "enzyme-to-json": "^3.3.4",
-    "file-loader": "^1.1.11",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.0.4",
+    "css-loader": "^3.2.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
+    "enzyme-to-json": "^3.4.0",
+    "eslint": "^6.4.0",
+    "file-loader": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^1.2.0",
+    "husky": "^2.7.0",
     "identity-obj-proxy": "^3.0.0",
     "immutable": "^4.0.0-rc.12",
-    "jest": "^22.4.4",
-    "jest-canvas-mock": "^2.1.0",
-    "jest-resolve": "^22.4.3",
-    "lint-staged": "^7.1.2",
-    "mini-css-extract-plugin": "^0.4.0",
-    "moment-timezone": "^0.5.23",
+    "jest": "^24.9.0",
+    "lint-staged": "^9.2.5",
+    "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "^4.12.0",
-    "optimize-css-assets-webpack-plugin": "^4.0.1",
-    "saas": "^1.0.0",
-    "sass-loader": "^7.0.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "sass-loader": "^7.1.0",
     "script-loader": "^0.7.2",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.21.0",
-    "stylelint": "^9.9.0",
-    "stylelint-config-sass-guidelines": "^5.0.0",
-    "stylelint-no-px": "^0.12.0",
-    "stylelint-order": "^1.0.0",
-    "stylelint-scss": "^3.1.0",
-    "ts-jest": "^22.4.6",
-    "ts-loader": "^3.5.0",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.16.0",
-    "tslint-config-airbnb": "^5.11.1",
+    "style-loader": "^1.0.0",
+    "stylelint": "^11.0.0",
+    "stylelint-config-sass-guidelines": "^6.1.0",
+    "stylelint-no-px": "^0.12.1",
+    "stylelint-order": "^3.1.1",
+    "stylelint-scss": "^3.11.1",
+    "sw-precache-webpack-plugin": "^0.11.5",
+    "terser-webpack-plugin": "^2.1.0",
+    "ts-jest": "^24.1.0",
+    "ts-loader": "^6.1.2",
+    "ts-node": "^8.4.1",
+    "tslint": "^5.20.0",
+    "tslint-config-airbnb": "^5.11.2",
     "tslint-eslint-rules": "^5.4.0",
     "tslint-jsdoc-rules": "^0.2.0",
-    "typescript": "^3.3.3",
-    "typescript-eslint-parser": "^21.0.1",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.3.1",
-    "webpack-merge": "^4.2.1",
+    "typescript-eslint-parser": "^22.0.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack": "^4.40.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.1",
+    "webpack-manifest-plugin": "^2.1.1",
     "webpack-obfuscator": "^0.18.0"
   },
   "lint-staged": {
