@@ -3,21 +3,16 @@ const dotenv = require('dotenv');
 
 // importing webpack dependencies
 const webpack = require('webpack');
-const cleanWebpackPlugin = require('clean-webpack-plugin');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const htmlWebpackPlugin = require('html-webpack-plugin');
 const miniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 
-const getFaviconUrl = () => {
-  return `https://res.cloudinary.com/mashafrancis/image/upload/v1551124424/kari4me/favicon.png`;
-};
-
 // instantiating webpack dependencies
-const cleanWebpack = new cleanWebpackPlugin();
+const cleanWebpack = new CleanWebpackPlugin();
 const htmlWebpack = new htmlWebpackPlugin({
   template: 'src/index.html',
   title: 'kari4me',
-  getFaviconUrl,
   minify: {
     removeComments: true,
     collapseWhitespace: true
