@@ -13,13 +13,16 @@ export interface TripsPageFormState {
   isLoading: boolean;
   isValid: boolean;
   focused: boolean;
+  fields: {
+    [key: string]: string | number | any
+  };
   locations: {
     origin: Location,
     destination: Location,
   };
   dates: {
     [key: string]: string | number
-  }
+  };
   errors: {
     [key: string]: string
   };
